refactor: 実依存を配線する 4 モジュールが命名で coverage の外に出る
`src/lib/auth/auth.ts`、`src/lib/auth/session.ts`、`src/lib/drizzle/db.ts`、
`tools/vite-plugins/wrangler-types-plugin.ts` を `*.live.ts` に改名し、
`vitest.config.mts` の `coverageExclude` からこの 4 パスを消した。
既存の `{src,scripts,tools}/**/*.live.ts` パタンが 4 本を拾う。

why: 4 本ともロジックは別モジュールにあり、自分では実依存
（Cloudflare binding、better-auth、node の spawn/stat）を配線するだけ。
パスで並んでいる限り、この種のモジュールが増えるたびに config の編集が要る。

`tools/oxlint-plugins/arch-rules.test.ts` の `it.each` が並べる
`@/lib/auth/auth` を `@/lib/auth/auth.live` に直した。同じ行の
`@/lib/auth/actions.live` を #163 が改名に合わせて直したのと同じ理由で、
行が名指すモジュールの実際の specifier に合わせる。
`arch-rules.js` の ban target は `resolveImportTarget` が `.live` を
落としてから照合するので編集していない。
